Adds agreement and policy mocks, together with some Gherkin tests
