+++ examples/mock/caddy/Caddyfile
@@ -13,8 +13,20 @@
   route @api {
     uri strip_prefix /api
 
+    # Static responses that take precedence over the Django application
+    handle /service/policy/1/ {
+        header "Content-Type" "application/json"
+        rewrite * "policy-1.json"
+        file_server
+    }
+    handle /service/agreement/1/ {
+        header "Content-Type" "application/json"
+        rewrite * "agreement-1-mcc.json"
+        file_server
+    }
+
     # An example of a static response, taking precedence over the reverse proxy to the Django application
-    handle /service/policy/ {
+    handle /service/agreement/1/ {
         header "Content-Type" "application/json"
         rewrite * "policy-1.json"
         file_server
